[#21] refactor : member 패키지 생성 리팩토링
