--- src/main/java/com/flab/goodchoice/coupon/application/MemberCommand.java
@@ -1,7 +1,0 @@
-package com.flab.goodchoice.coupon.application;
-
-import com.flab.goodchoice.coupon.domain.Member;
-
-public interface MemberCommand {
-    Member save(Member member);
-}
