+++ src/main/java/com/flab/goodchoice/coupon/infrastructure/CouponPublishCommandImpl.java
@@ -4,7 +4,7 @@ import com.flab.goodchoice.coupon.application.CouponPublishCommand;
 import com.flab.goodchoice.coupon.domain.CouponPublish;
 import com.flab.goodchoice.coupon.infrastructure.entity.CouponEntity;
 import com.flab.goodchoice.coupon.infrastructure.entity.CouponPublishEntity;
-import com.flab.goodchoice.coupon.infrastructure.entity.MemberEntity;
+import com.flab.goodchoice.member.infrastructure.entity.MemberEntity;
 import com.flab.goodchoice.coupon.infrastructure.repositories.CouponPublishRepository;
 import org.springframework.stereotype.Component;
