+++ src/main/java/com/flab/goodchoice/coupon/infrastructure/entity/CouponPublishEntity.java
@@ -1,6 +1,7 @@
 package com.flab.goodchoice.coupon.infrastructure.entity;
 
 import com.flab.goodchoice.coupon.domain.CouponPublish;
+import com.flab.goodchoice.member.infrastructure.entity.MemberEntity;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
