+++ src/main/java/com/flab/goodchoice/member/application/MemberCommand.java
@@ -1,0 +1,7 @@
+package com.flab.goodchoice.member.application;
+
+import com.flab.goodchoice.member.domain.model.Member;
+
+public interface MemberCommand {
+    Member save(Member member);
+}
