+++ src/main/java/com/flab/goodchoice/member/application/MemberQuery.java
@@ -1,0 +1,7 @@
+package com.flab.goodchoice.member.application;
+
+import com.flab.goodchoice.member.domain.model.Member;
+
+public interface MemberQuery {
+    Member findById(Long memberId);
+}
