+++ src/main/java/com/flab/goodchoice/member/infrastructure/MemberQueryImpl.java
@@ -1,11 +1,11 @@
-package com.flab.goodchoice.coupon.infrastructure;
+package com.flab.goodchoice.member.infrastructure;
 
-import com.flab.goodchoice.coupon.application.MemberQuery;
-import com.flab.goodchoice.coupon.domain.Member;
-import com.flab.goodchoice.coupon.exception.MemberError;
-import com.flab.goodchoice.coupon.exception.MemberException;
-import com.flab.goodchoice.coupon.infrastructure.entity.MemberEntity;
-import com.flab.goodchoice.coupon.infrastructure.repositories.MemberRepository;
+import com.flab.goodchoice.member.application.MemberQuery;
+import com.flab.goodchoice.member.domain.model.Member;
+import com.flab.goodchoice.member.exception.MemberError;
+import com.flab.goodchoice.member.exception.MemberException;
+import com.flab.goodchoice.member.infrastructure.entity.MemberEntity;
+import com.flab.goodchoice.member.domain.repositories.MemberRepository;
 import org.springframework.stereotype.Component;
 
 @Component
