+++ src/test/java/com/flab/goodchoice/coupon/infrastructure/FakeMemberQuery.java
@@ -1,11 +1,11 @@
 package com.flab.goodchoice.coupon.infrastructure;
 
-import com.flab.goodchoice.coupon.application.MemberQuery;
-import com.flab.goodchoice.coupon.domain.Member;
-import com.flab.goodchoice.coupon.exception.MemberError;
-import com.flab.goodchoice.coupon.exception.MemberException;
-import com.flab.goodchoice.coupon.infrastructure.entity.MemberEntity;
-import com.flab.goodchoice.coupon.infrastructure.repositories.MemberRepository;
+import com.flab.goodchoice.member.application.MemberQuery;
+import com.flab.goodchoice.member.domain.model.Member;
+import com.flab.goodchoice.member.exception.MemberError;
+import com.flab.goodchoice.member.exception.MemberException;
+import com.flab.goodchoice.member.infrastructure.entity.MemberEntity;
+import com.flab.goodchoice.member.domain.repositories.MemberRepository;
 
 public class FakeMemberQuery implements MemberQuery {
